requirements.txt
hikari
hikari-lightbulb
openai